--- src/KycDaoClient.css
@@ -1,41 +1,0 @@
-.KycDaoModalFrame {
-	box-shadow: 0px 1px 174px rgba(59, 31, 69, 0.2);
-	width: 100%;
-	height: 100%;
-	background-color: #fefefe;
-	position: absolute;
-	overflow: hidden;
-}
-
-.KycDaoModalRoot {
-	width: 100%;
-	position: absolute;
-	inset: 0;
-}
-
-@media only screen and (min-width: 992px) {
-	.KycDaoModalFrame {
-		width: var(--width);
-		height: var(--height);
-		margin: auto;
-		border-radius: 12px;
-		overflow: hidden;
-		inset: 0;
-	}
-
-	.KycDaoModalRoot {
-		position: sticky;
-		overflow: hidden;
-		height: 100vh;
-	}
-}
-
-@media only screen and (max-width: 992px) {
-	body:has(.KycDaoModalRoot) {
-		overflow: hidden;
-	}
-}
-
-body:has(.KycDaoModalRoot) {
-	background: var(--kyc-dao-backdrop, auto);
-}
