--- src/KycDaoIframeClient.css
@@ -1,27 +1,0 @@
-.KycDaoModalIFrame {
-	box-shadow: 0px 1px 174px rgba(59, 31, 69, 0.2);
-	width: 100%;
-	height: 100%;
-	background-color: #fefefe;
-	position: absolute;
-	overflow: hidden;
-
-	border: 0px;
-	border-width: 0;
-}
-
-@media only screen and (min-width: 992px) {
-	.KycDaoModalIFrame {
-		width: var(--width);
-		height: var(--height);
-		margin: auto;
-		border-radius: 12px;
-	}
-}
-
-.KycDaoModalRoot {
-	background: var(--backdrop, auto);
-	position: absolute;
-	width: 100%;
-	inset: 0;
-}
