+++ src/StandaloneClient.ts
@@ -15,7 +15,7 @@ import {
 	messageHndlr,
 	nearNetworkRegex,
 	nearRedirectCheck,
-} from "./KycDaoClientCommon"
+} from "./StandaloneClientCommon"
 
 // basically the KycDaoClient.css
 const styles = `
