+++ src/index.js
@@ -1,5 +1,5 @@
 export * from "./BootstrapKycDaoModal"
 export * from "./App"
 export * from "./widget"
-export * from "./KycDaoClient"
-export * from "./KycDaoIframeClient"
+export * from "./StandaloneClient"
+export * from "./StandaloneIframeClient"
